formatting

+++ src/utils/common_utils/utils.py
@@ -1,10 +1,12 @@
 import pandas as pd
 import os
-#import wget
 import requests
 import matplotlib.pyplot as plt
 
 import streamlit as st
+import folium
+from streamlit_folium import folium_static
+
 
 def download_usgs_data(site_id, begin_year):
     """
@@ -19,19 +21,25 @@ def download_usgs_data(site_id, begin_year):
     try:
         yesterday = pd.Timestamp.now() - pd.Timedelta(days=1)
         yesterday_str = yesterday.strftime('%Y-%m-%d')
-        
+
+        information_url = f"https://waterdata.usgs.gov/nwis/inventory/?site_no={site_id}&agency_cd=USGS"
         url = f"https://waterdata.usgs.gov/nwis/dv?cb_00060=on&format=rdb&site_no={site_id}&legacy=&referred_module=sw&period=&begin_date={begin_year}-01-01&end_date={yesterday_str}"
+
         if not os.path.exists("data/temp"):
             os.makedirs("data/temp")
         file_path = requests.get(url)
+        info_path = requests.get(information_url)
         #save the file to the temp directory
         with open(os.path.join("data/temp","flow_data.txt"), 'wb') as f:
             f.write(file_path.content)
+        with open(os.path.join("data/temp","info_data.txt"), 'wb') as f:
+            f.write(info_path.content)
 
         file_path = os.path.join("data/temp","flow_data.txt")
+        info_path = os.path.join("data/temp","info_data.txt")
         
         
-        return file_path
+        return file_path, info_path
     except Exception as e:
         st.error(f"Error downloading peak flow data: {e}")
        
@@ -63,13 +71,90 @@ def load_flow_data(file_path):
         df.loc[(df["month"] < 4) | (df["month"] == 12), "season"] = "Winter"
         #add seasons to the dataframe
         
-        os.remove(file_path)  # Remove the downloaded file after loading
+        
         return df
     except Exception as e:
         st.error(f"Error loading peak flow data: {e}")
        
         return None
     
+
+
+
+
+
+def clean_temp_files(file_path, info_path):
+    """
+    Cleans up temporary files created during the download process.
+    
+    Args:
+        file_path (str): The path to the downloaded file.
+        info_path (str): The path to the info file.
+    """
+    try:
+        if os.path.exists(file_path):
+            os.remove(file_path)
+        if os.path.exists(info_path):
+            os.remove(info_path)
+    except Exception as e:
+        st.error(f"Error cleaning up temporary files: {e}")
+
+def extract_site_info(info_path):
+    """
+    Extracts site information from the info file.
+    
+    Args:
+        info_path (str): The path to the info file.
+        
+    Returns:
+        dict: A dictionary containing site information.
+    """
+    #try:
+    with open(info_path, 'r') as f:
+        lines = f.readlines()
+    site_info = {}
+    for line in lines:
+        #find lines that contain latitude and longitude
+        if "latitude" in line.lower() or "longitude" in line.lower():
+            parts = line.strip().split(' ')
+            for part in parts:
+                if part == "" or part == " ":
+                    parts.remove(part)
+            
+            
+            latitude = parts[1]
+            latitude = latitude.replace(',', ' ').replace('&#176', ' ').replace(';', ' ').replace("'", ' ').replace('"', ' ')
+            
+            latitude = latitude.split(" ")
+            
+            decimal_lat = float(latitude[0]) + (float(latitude[2]) / 60) + (float(latitude[3])/ 3600)
+            longitude = parts[4]
+            longitude = longitude.replace(',', ' ').replace('&#176', ' ').replace(';', ' ').replace("'", ' ').replace('"', ' ')
+            longitude = longitude.split(" ")
+            
+            decimal_long = float(longitude[0]) + (float(longitude[2]) / 60) + (float(longitude[3])/ 3600)
+            location_df = pd.DataFrame({'latitude': [decimal_lat], 'longitude': [decimal_long*-1]})
+            return location_df
+            
+                
+def create_location_plot(info_path, site_id):
+    location_df = extract_site_info(info_path)
+    attr = (
+    '&copy; <a href="https://www.stadiamaps.com/" target="_blank">Stadia Maps</a> &copy; <a href="https://openmaptiles.org/" target="_blank">OpenMapTiles</a> &copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> '
+    'contributors, &copy; <a href="https://cartodb.com/attributions">CartoDB</a>'
+)
+    tiles = 'https://tiles.stadiamaps.com/tiles/alidade_satellite/{z}/{x}/{y}{r}.jpg'
+    
+    m = folium.Map(location=[location_df["latitude"],location_df["longitude"]], tiles=tiles,attr = "Aerial Imagery", zoom_start=16)
+    
+    folium.Marker(
+        [location_df["latitude"], location_df["longitude"]], popup=f"Gage {site_id} location", tooltip=f"Gage {site_id} location"
+    ).add_to(m)
+    
+    #folium.LayerControl().add_to(m)
+    st.header(f"Gage {site_id} Location")
+    folium_static(m, width=1800, height=500)
+            
 
 def subset_by_season(df):
     spring_dict = {}
